refactor(api): drop createAndSubmitProposal helper in favor of createProposal status arg (#1041)
refactor(api): drop createAndSubmitProposal helper in favor of
createProposal status arg
refactor(api): inline pipelineTestFixtures into TestDecisionsDataManager
