+++ services/api/src/routers/decision/instances/getProposalsForPhase.test.ts
@@ -1,6 +1,8 @@
 import { getProposalsForPhase } from '@op/common';
 import { db, eq, sql } from '@op/db/client';
 import {
+  ProcessStatus,
+  ProposalStatus,
   processInstances,
   proposals,
   stateTransitionHistory,
@@ -9,12 +11,9 @@ import { describe, expect, it } from 'vitest';
 
 import { TestDecisionsDataManager } from '../../../test/helpers/TestDecisionsDataManager';
 import {
-  createAndSubmitProposal,
-  createInstanceWithSchema,
-  executeTestTransition,
   schemaWithPipeline,
   schemaWithoutPipeline,
-} from '../../../test/helpers/pipelineTestFixtures';
+} from '../../../test/helpers/pipelineSchemas';
 
 // Random UUID unlikely to match any real transition
 const NONEXISTENT_PHASE_ID = '00000000-0000-0000-0000-000000000000';
@@ -25,22 +24,26 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal 1 ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal 2 ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
 
@@ -57,22 +60,25 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Create and submit 3 proposals; pipeline limits to 2
     for (let i = 1; i <= 3; i++) {
-      await createAndSubmitProposal(testData, caller, {
+      await testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal ${i} ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -88,22 +94,26 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Active ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Soft-deleted ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
 
@@ -124,22 +134,25 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Create and submit 3 proposals; pipeline limits to 2
     for (let i = 1; i <= 3; i++) {
-      await createAndSubmitProposal(testData, caller, {
+      await testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal ${i} ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -159,11 +172,12 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
 
     const result = await getProposalsForPhase({
       instanceId,
@@ -178,27 +192,31 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Active after transition ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Soft-deleted after transition ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
 
     // Both proposals survive the transition (no pipeline)
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -221,17 +239,20 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     // Create one submitted proposal and one draft (not submitted)
-    const submitted = await createAndSubmitProposal(testData, caller, {
+    const submitted = await testData.createProposal({
       userEmail,
       processInstanceId: instanceId,
       proposalData: { title: `Submitted ${task.id}` },
+      status: ProposalStatus.SUBMITTED,
     });
     await testData.createProposal({
       userEmail,
@@ -251,22 +272,26 @@ describe.concurrent('getProposalsForPhase', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Legacy proposal 1 ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Legacy proposal 2 ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
