+++ services/api/src/routers/decision/proposals/listProposalsPhaseScoped.test.ts
@@ -1,15 +1,13 @@
 import { db, eq } from '@op/db/client';
-import { proposals } from '@op/db/schema';
+import { ProcessStatus, ProposalStatus, proposals } from '@op/db/schema';
 import { describe, expect, it } from 'vitest';
 
 import { TestDecisionsDataManager } from '../../../test/helpers/TestDecisionsDataManager';
 import {
-  createAndSubmitProposal,
-  createInstanceWithSchema,
-  executeTestTransition,
   schemaWithPipeline,
   schemaWithoutPipeline,
-} from '../../../test/helpers/pipelineTestFixtures';
+} from '../../../test/helpers/pipelineSchemas';
+import { createAuthenticatedCaller } from '../../../test/supabase-utils';
 
 describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
   it('returns only selected proposals after a transition with a limiting pipeline', async ({
@@ -17,22 +15,26 @@ describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
+    const caller = await createAuthenticatedCaller(userEmail);
 
     // Create and submit 3 proposals; the pipeline limits to 2
     for (let i = 1; i <= 3; i++) {
-      await createAndSubmitProposal(testData, caller, {
+      await testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal ${i} ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -51,21 +53,25 @@ describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
+    const caller = await createAuthenticatedCaller(userEmail);
 
     for (let i = 1; i <= 3; i++) {
-      await createAndSubmitProposal(testData, caller, {
+      await testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Proposal ${i} ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       });
     }
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -84,22 +90,27 @@ describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
+    const caller = await createAuthenticatedCaller(userEmail);
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Active proposal ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `To-be-deleted proposal ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
 
@@ -109,7 +120,7 @@ describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
       .set({ deletedAt: new Date().toISOString() })
       .where(eq(proposals.id, p2.id));
 
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
@@ -129,27 +140,32 @@ describe.concurrent('listProposals: phase-scoped proposal visibility', () => {
     onTestFinished,
   }) => {
     const testData = new TestDecisionsDataManager(task.id, onTestFinished);
-    const { instanceId, userEmail, caller } = await createInstanceWithSchema(
-      testData,
-      task.id,
-      schemaWithoutPipeline,
-    );
+    const setup = await testData.createDecisionSetup({
+      processSchema: schemaWithoutPipeline,
+      instanceCount: 1,
+      status: ProcessStatus.PUBLISHED,
+    });
+    const instanceId = setup.instances[0]!.instance.id;
+    const { userEmail } = setup;
+    const caller = await createAuthenticatedCaller(userEmail);
 
     const [p1, p2] = await Promise.all([
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `Active proposal ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
-      createAndSubmitProposal(testData, caller, {
+      testData.createProposal({
         userEmail,
         processInstanceId: instanceId,
         proposalData: { title: `To-be-deleted after transition ${task.id}` },
+        status: ProposalStatus.SUBMITTED,
       }),
     ]);
 
     // Transition first (both proposals make it into the join table)
-    await executeTestTransition({
+    await testData.advancePhase({
       instanceId,
       fromPhaseId: 'submission',
       toPhaseId: 'review',
